Add Offcanvas for controls, move infrastructure components, and add MLEx icon
