+++ src/assets/lse-style.css
@@ -37,3 +37,16 @@ label {
     font-size: large;
     font-weight: bold;
 }
+
+.offcanvas-header {
+    display: none;
+}
+
+.offcanvas-body {
+    padding: 0%;
+}
+
+.navbar {
+    z-index: 1100 !important;
+    position: relative;
+}
